docs: update `.gitignore` to exclude media files
Grouped all demo-generated media (e.g., images and videos)
into a local `media/` folder, which is now excluded from Git tracking.
This improves the organization of the `RayTracer` directory, making it more concise.
The structure of the `media/` folder is left up to the user.

---

**Note:** The creation of the `media/` folder is not strictly necessary,
as file types like `.mp4`, `.png`, and `.pfm` are already ignored.
This commit simply clarifies the rationale for explicitly excluding the `media/` folder.

---