fix: remove clamp from `write_ldr_image`
A clamp was added before gamma correction in `write_ldr_image` to correct high pixel values in the demo, observed particularly when using an `ImagePigment`. These high values caused issues because no normalization or clamping was applied in `demo.jl`, as the `.png` output was intended only as a quick preview without customizable tone mapping.

There is no need to clamp values directly in `write_ldr_image`, the correct usage is to normalize and clamp before writing ldr image. This commit reverts that change, restoring the original behavior of `write_ldr_image`, and instead applies tone mapping with default parameters directly in `demo.jl`.

Note: This issue would not occur when creating the main program, where tone mapping is always performed because parameters could be specified by the user.