add script.jl as central handler using Comonic framework
Created script.jl intended to unify all functionalities in the codebase under Comonic.
This initial version is not fully tested and requires improved exception handling.

Shared to serve as a foundation for expanding features from past releases such as pfm2image, tonemapping, flattracer, and onofftracer.

For each functionality, a @cast function command() should be created following the example of pathtracer.